Update p5-Sys-SigAction-Changes to 0.15.
Changes from previous:
0.15  1 Jul 2011

Clean up POD. Close bug #69057. Other minor tweaks to POD.

0.13  23 Jun 2011

No functional changes. Fix for test timeout.t.
Fix strict undefined symbol error in timeout.t, when Time::HiRes is not present.
Not sure if constant pragma will exist in all supported perl versions,
so, we just commented out the use strict in this test.

Print warning when Time::HiRes not found in Makefile.PL

0.12  20 Jun 2011

Conditionally add 'LICENSE' => 'perl' to WriteMakefile()
call if $ExtUtils::MakeMaker::VERSION >= 6.3002.

Added support for timeout_call() in fractional seconds
expressed as a floating point number.  If Time::HiRes
is not loadable, then the timeout value is raised to the
next high integer value with the POSIX:ceil() funtion.

Added sig_alarm(), which timeout_call uses. This is drop
in replacement for alarm(). If Time::HiRes
is not loadable, then the seconds argument is raised to the
next high integer value with the POSIX:ceil() funtion.