Import lettuce 1.1.33
Lettuce is a simple behaviour-driven development (BDD) tool based
on Cucumber. The features of the system (to be tested) are defined
in a natural language with scenarios describing the features. Steps
are written in python.

(I packaged this for my work at ISC. We are developing a
DNS server test suite to be usable for various DNS servers
for compliance testing.)