General cleanups.
The old starship site is gone.  Get from NetBSD.org now.

Change homepage to a working one.

Add license.  Note that the only license reference I could find was "It
is distributed for free under a BSD-style license" which I took to mean
"original-bsd."

Doesn't work under Python 3.  Since it is deprecated in favour of the
py-imaging module there is no point in struggling with it.

Added comment to patch and regenerated distinfo.

No PKGREVISION bump since there are no changes to any code that ever built.