editors/xemacs: Fix all non-NetBSD platforms
The previous change to PLIST broke all platforms not NetBSD if built under
PKG_DEVELOPER=yes.  The netbsd.h header will be present on all platforms,
so it needs to have a dedicated entry in PLIST (as it used to).  The
LOWER_OPSYS variable is just wrong in this case.

Secondly, the dragonfly.h file was updated for the default compiler,
allowing this version of xemacs to build again.