mail/Pantomine: Fix timezone redefinition for DragonFly
Pantomine defines a structure named "timezone" which clashes with the
system definition.  It was renamed to "timezonePantomine".  While
technically it may result in change in the pkg binary, no PKGREVISION
bump is necessary because it's functionally the same as the previous
version, assuming it actually built on the platform in question.