Update to 1.8.9.3:
pkgsrc changes:

- maintain patch naming convention
- detect kqueue properly

The Asterisk Development Team has announced the release of Asterisk 1.8.9.3.

The release of Asterisk 1.8.9.3 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following are the issues resolved in this release:

* --- Fix ACK routing for non-2xx responses.
  (Closes issue ASTERISK-19389. Reported by: Karsten Wemheuer)

* --- Fix regressions with regards to route-set creation on early dialogs ---
  (Closes issue ASTERISK-19358. Reported-by: Karsten Wemheuer)

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-1.8.9.3

Thank you for your continued support of Asterisk!