Added filesystems/cloudfuse version 20120124
Filesystem (fuse) implemented on Mosso's Cloud Files

Cloudfuse is a FUSE application which provides access to Rackspace's
Cloud Files (or any installation of Swift).

Cloud Files is a remote storage system which is similar in principle to
Amazon S3.  It provides a simple RESTful interface to storing and retrieving
objects.

http://www.rackspacecloud.com/cloud_hosting_products/files